Clean-up nav architecture
