+++ pages/index.js
@@ -1,7 +1,7 @@
 import useStore from '../hooks/useStore';
 import ProductCard from '../components/ProductCard';
-import Nav from '../components/MainNav';
-import Header from '../components/Header'
+import IndexNav from '../components/IndexNav';
+import Header from '../components/Header';
 
 export default function Home() {
   const products = useStore(state => state.products);
@@ -10,9 +10,9 @@ export default function Home() {
     <>
       {products.map(product => (
         <ProductCard key={product.id} product={product} />
-        ))}
-        <Header/>
-      <Nav/>
+      ))}
+      <Header />
+      <IndexNav />
     </>
   );
 }
